+++ src/JsonApi/Schema/RelativeLinks.php
@@ -1,0 +1,15 @@
+<?php
+namespace Src\Utils\Hateoas;
+
+use WoohooLabs\Yin\JsonApi\Schema\Links;
+
+class RelativeLinks extends Links
+{
+    /**
+     * @param array $links
+     */
+    public function __construct(array $links)
+    {
+        parent::__construct(gethostname(), $links);
+    }
+}